Move OrderConfirmationEmail into Order namespace
